Support native Blaze stages with compact runtime args and bounded host compilation

+++ tt_metal/api/tt-metalium/program_descriptors.hpp
@@ -60,7 +60,8 @@ struct TileDescriptor {
 };
 
 struct CBFormatDescriptor {
-    uint8_t buffer_index = 0;
+    // Host-side logical IDs may exceed the hardware range before phase remapping.
+    uint16_t buffer_index = 0;
     tt::DataFormat data_format = tt::DataFormat::Float32;
     uint32_t page_size = 0;
     std::optional<TileDescriptor> tile;
